requirements.txt
fastmcp>=0.1.0
xrpl-py>=2.6.0
python-dotenv>=1.0.0